Spruced up ModManagementScreen - phase 1 - patch1
